tkinter-python-ui/requirements.txt
numpy==1.15.0
opencv-python==3.4.2.17
Pillow==5.2.0